+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/Binary/RoundTrip.hs
@@ -1,8 +1,10 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
+{-# LANGUAGE DataKinds #-}
 {-# LANGUAGE FlexibleContexts #-}
-{-# LANGUAGE MonoLocalBinds #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# OPTIONS_GHC -Wno-orphans #-}
 
 module Test.Cardano.Ledger.Shelley.Binary.RoundTrip (
   roundTripShelleyCommonSpec,
@@ -12,6 +14,8 @@ module Test.Cardano.Ledger.Shelley.Binary.RoundTrip (
 import Cardano.Ledger.Binary
 import Cardano.Ledger.Compactible
 import Cardano.Ledger.Core
+import Cardano.Ledger.Crypto (Crypto)
+import Cardano.Ledger.Shelley (ShelleyEra)
 import Cardano.Ledger.Shelley.Governance
 import Cardano.Ledger.Shelley.LedgerState
 import Test.Cardano.Ledger.Common
@@ -39,11 +43,13 @@ roundTripShelleyCommonSpec ::
   , Arbitrary (GovState era)
   , Arbitrary (PParams era)
   , Arbitrary (PParamsUpdate era)
+  , RuleListEra era
   ) =>
   Spec
 roundTripShelleyCommonSpec = do
   roundTripCoreEraTypesSpec @era
   roundTripStateEraTypesSpec @era
+  roundTripAllPredicateFailures @era
 
 roundTripStateEraTypesSpec ::
   forall era.
@@ -67,3 +73,17 @@ roundTripStateEraTypesSpec = do
     roundTripShareEraTypeSpec @era @UTxOState
     roundTripEraTypeSpec @era @EpochState
     roundTripEraTypeSpec @era @NewEpochState
+
+instance Crypto c => RuleListEra (ShelleyEra c) where
+  type
+    EraRules (ShelleyEra c) =
+      '[ "DELEG"
+       , "DELEGS"
+       , "DELPL"
+       , "LEDGER"
+       , "LEDGERS"
+       , "POOL"
+       , "PPUP"
+       , "UTXO"
+       , "UTXOW"
+       ]