+++ libs/cardano-ledger-core/CHANGELOG.md
@@ -31,6 +31,7 @@
 
 ### `testlib`
 
+* Add `EraRuleProof`, `UnliftRules`, `roundTripAllPredicateFailures`
 * Add `Test.Cardano.Ledger.Plutus.ScriptTestContext`
 * Add `genByronVKeyAddr`, `genByronAddrFromVKey`
 * Add `Uniform` instances for `ByronKeyPair` and `KeyPair`